@@ -7,7 +7,7 @@
 package com.qsboy.antirecall.ui.widget;
 
 import android.content.Context;
-import android.support.v7.widget.SwitchCompat;
+import androidx.appcompat.widget.SwitchCompat;
 import android.util.AttributeSet;
 
 import java.lang.reflect.Field;
